product-service/requirements.txt
django>=6.0
djangorestframework
djangorestframework-simplejwt
psycopg[binary]>=3.1.12
requests
